Add layout files for new "slides" post type
The files being added in this commit add support for a new "slides" type
of post, which will use "remarkjs" to convert the markdown text of the
post into traditional presentation slides. Simply add the following:

    type = "slides"

to the front matter of the post to make use of this new layout (i.e.
alongside the normal "title", "date", etc. that's normally present).